requirements.txt
boto3==1.35.74
Flask==3.1.0
keyboard==0.13.5
langchain==0.3.11
langchain_aws==0.2.9
langchain_core==0.3.24
langchain_openai==0.2.12
langgraph==0.2.57
milvus_model==0.2.5
msgpack_python==0.5.6
openai==1.57.2
pandas==2.2.3
pyautogui==0.9.54
pydantic==2.10.3
pymilvus==2.4.9
pymssql==2.3.1
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.0
SQLAlchemy==2.0.34
transformers==4.44.2
twilio==8.3.0
typing_extensions==4.12.2
langgraph-checkpoint==2.0.4
langgraph-checkpoint-sqlite==2.0.1
pyodbc==5.1.0
transformers==4.44.2
sentence-transformers==3.2.1
openpyxl==3.1.5
python-dotenv==1.0.1
streamlit==1.31.1
pypdf==4.2.0
cryptography==42.0.5